fix: Better DOMready() and simplified decimals
The window load event is only fired when the entire page has been
loaded, including all resourses like scripts, stylesheets and images.
This is unreliable especially with ad- and privacy blockers preventing
certain files to load. Replaced it with a DOMready() function that
triggers earlier and always.

The decimals handling of prices has been simplified using toFixed(2).

Note on Number.EPSILON; It doesn't prevent floating-point errors in
the way it's used here. In rare occasions it could prevent incorrectly
rounding down instead of up but the 0.01 difference doesn't matter
here and we're also not doing direct equality comparisons.