feat: Changes how prices are calculated
Prices for each item are now calculated using the lowest Arx package
possible to buy that specific item. It does not make sense if an item
costs 12500 to use the 8400 package and calculate leftovers with the
5000 package. This would be more expensive than simply buying the
larger 16800 package.

Also a few fixes I missed and all my tabs were messed up...